backend/requirements.txt
asgiref
Django==3.2.12
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
mysqlclient==2.1.0
Pillow